% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitiveR.R
\docType{package}
\name{cognitiveR}
\alias{cognitiveR}
\alias{cognitiveR-package}
\title{cognitiveR: An R Package of Microsoft Cognitive API.}
\description{
The cognitiveR package provides a series of functions that allow R
users to access Microsoft Cognitive REST APIs
\url{https://www.microsoft.com/cognitive-services},
which includes machine learning models of vision, speech, language.
}

