% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitive.R
\name{cognitiveSentiAnalysis}
\alias{cognitiveSentiAnalysis}
\title{Sentiment analysis.}
\usage{
cognitiveSentiAnalysis(text, apiKey)
}
\arguments{
\item{text}{Text to be analyzed. Can be a dataframe, list, or text vector.}

\item{apiKey}{API key.}
}
\value{
Response from sentiment analysis API, including score, text id, and errors if there are.
}
\note{
See \url{https://westus.dev.cognitive.microsoft.com/docs/services/TextAnalytics.V2.0} for more information about Text Analytics on Microsoft Cognitive Analysis.
}

