% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectLanguage.R
\name{detectLanguage.buildOneDocument}
\alias{detectLanguage.buildOneDocument}
\title{detectLanguage.buildOneDocument}
\usage{
detectLanguage.buildOneDocument(id = "", text = "nice!")
}
\description{
https://westus.dev.cognitive.microsoft.com/docs/services/TextAnalytics.V2.0/operations/56f30ceeeda5650db055a3c7
}
\examples{
detectLanguage.buildOneDocument()
}

