% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectLanguage.R
\name{detectLanguage.buildQueryBody}
\alias{detectLanguage.buildQueryBody}
\title{buildQueryBody}
\usage{
detectLanguage.buildQueryBody(...)
}
\description{
https://westus.dev.cognitive.microsoft.com/docs/services/TextAnalytics.V2.0/operations/56f30ceeeda5650db055a3c7
}
\examples{
detectLanguage.buildQueryBody(detectLanguage.buildOneDocument(),detectLanguage.buildOneDocument())
}

