% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectLanguage.R
\name{detectLanguage.getLanguages}
\alias{detectLanguage.getLanguages}
\title{detectLanguage.getLanguages}
\usage{
detectLanguage.getLanguages(apiKey, ...)
}
\description{
https://westus.dev.cognitive.microsoft.com/docs/services/TextAnalytics.V2.0/operations/56f30ceeeda5650db055a3c7
}
\examples{
detectLanguage.getLanguages("yourApiKey", list(id="", text= "nice, it's perfect!"),list(id="", text= "今天天氣很好"),list(id="", text= "ཀོང་དགས་།"))
}

