% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyPhrases.R
\name{keyPhrases.buildOneDocument}
\alias{keyPhrases.buildOneDocument}
\title{keyPhrases.buildOneDocument}
\usage{
keyPhrases.buildOneDocument(lang = "en", id = "",
  text = "nice! it's perfect!")
}
\description{
https://westus.dev.cognitive.microsoft.com/docs/services/TextAnalytics.V2.0/operations/56f30ceeeda5650db055a3c7
}
\examples{
keyPhrases.buildOneDocument()
}

