% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAnalytics_KeyPhrases.R
\name{keyPhrases.buildQueryBody}
\alias{keyPhrases.buildQueryBody}
\title{keyPhrases.buildQueryBody}
\usage{
keyPhrases.buildQueryBody(...)
}
\description{
https://westus.dev.cognitive.microsoft.com/docs/services/TextAnalytics.V2.0/operations/56f30ceeeda5650db055a3c7
}
\examples{
keyPhrases.buildQueryBody(keyPhrases.buildOneDocument(),keyPhrases.buildOneDocument())
}

