% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyPhrases.R
\name{keyPhrases.getKeyPhrases}
\alias{keyPhrases.getKeyPhrases}
\title{keyPhrases.getKeyPhrases}
\usage{
keyPhrases.getKeyPhrases(apiKey, ...)
}
\description{
https://westus.dev.cognitive.microsoft.com/docs/services/TextAnalytics.V2.0/operations/56f30ceeeda5650db055a3c7
}
\examples{
keyPhrases.getKeyPhrases("yourApiKey", list(lang="zh_cht", id="", text= "今天天氣很好"),list(lang="en", id="", text= "Very  bad hotel, room was very bad smells,  shower was very bad from now I never go to westbetern hotels."))
}

