% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_api_key}
\alias{load_api_key}
\title{Import API keys}
\usage{
load_api_key(path = "~/cognitive_api_keys.txt")
}
\arguments{
\item{path}{Path to API keys (yaml format). See Details.}
}
\description{
Import API keys
}
\details{
Default path to API keys file is ~/cognitive_api_keys.txt.

\preformatted{
Text-Analytics: xxxxxx
Face: xxxxxxx
}
}

