% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAnalytics_Sentiment.R
\name{sentiment.buildOneDocument}
\alias{sentiment.buildOneDocument}
\title{sentiment.buildOneDocument}
\usage{
sentiment.buildOneDocument(lang = "en", id = "", text = "nice!")
}
\description{
https://westus.dev.cognitive.microsoft.com/docs/services/TextAnalytics.V2.0/operations/56f30ceeeda5650db055a3c7
}
\examples{
sentiment.buildOneDocument()
}

