% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAnalytics_DetectLanguage.R
\name{text_detect_language}
\alias{text_detect_language}
\title{Text Analytics: Detect Language}
\usage{
text_detect_language(apiKey, ..., numberOfLanguagesToDetect = 1L)
}
\description{
Text Analytics: Detect Language
}
\examples{
\dontrun{
text_detect_language(load_api_key(),
                  list(text= "nice, it's perfect!"),
                  list(text= "今天天氣很好"),
                  list(text= "ཀོང་དགས་།"))
}
}
\references{
\url{https://westus.dev.cognitive.microsoft.com/docs/services/TextAnalytics.V2.0/operations/56f30ceeeda5650db055a3c7}
}

