% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAnalytics_Sentiment.R
\name{text_sentiment}
\alias{text_sentiment}
\title{Text Analytics: Sentiment}
\usage{
text_sentiment(apiKey, ...)
}
\description{
Text Analytics: Sentiment
}
\examples{
text_sentiment("yourApiKey", list(lang="en", id="", text= "nice!"),list(lang="en", id="", text= "nice!"),list(lang="en", id="", text= "nice!"))
}
\references{
\url{https://westus.dev.cognitive.microsoft.com/docs/services/TextAnalytics.V2.0/operations/56f30ceeeda5650db055a3c7}
}

