% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebML.R
\name{webML.breakIntoWords}
\alias{webML.breakIntoWords}
\title{webML.breakIntoWords}
\usage{
webML.breakIntoWords(apiKey, model, text, order = 5,
  maxNumOfCandidatesReturned = 5)
}
\arguments{
\item{apiKey}{string Subscription key which provides access to this API. Found in your subscriptions.}

\item{model}{string Which model to use, supported value: title/anchor/query/body}

\item{text}{string The line of text to break into words. If spaces are present, they will be interpreted as hard breaks and maintained, except for leading or trailing spaces, which will be trimmed.}

\item{order}{(optional) number The order of N-gram. If not specified, use default value 5 .Supported value: 1, 2, 3, 4, 5.}

\item{maxNumOfCandidatesReturned}{(optional) number Max number of candidates would be returned. If not specified, use default value 5.}
}
\description{
Break Into Words
Insert spaces into a string of words lacking spaces, like a hashtag or part of a URL. Punctuation or exotic characters can prevent a string from being broken, so it’s best to limit input strings to lower-case, alpha-numeric characters.
}
\details{
Ref: https://dev.projectoxford.ai/docs/services/55de9ca4e597ed1fd4e2f104/operations/565bf87b778daf12447f43c1
}
\examples{
webML.breakIntoWords(apiKey=""model = "body",text = "今天天氣很好")
}

