% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebML.R
\name{webML.lsModels}
\alias{webML.lsModels}
\title{webML.lsModels}
\usage{
webML.lsModels(apiKey)
}
\description{
https://dev.projectoxford.ai/docs/services/55de9ca4e597ed1fd4e2f104/operations/565bf87b778daf12447f43c1
}
\examples{
webML.lsModels(apiKey="")
}

