% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitive.R
\name{cognitiveTranslation}
\alias{cognitiveTranslation}
\title{Translation.}
\usage{
cognitiveTranslation(text, lanTo, apiKey)
}
\arguments{
\item{text}{text to be analyzed.}

\item{lanTo}{target language.}

\item{apiKey}{API key}

\item{lan_from}{source language.}
}
\value{
A data frame containing the translated text in the target language.
}
\note{
See \url{http://docs.microsofttranslator.com/text-translate.html#!/default/post_TranslateArray} for more information about Translation API.
}

